recipes/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Redis cache
redis==5.1.1

# Scraping
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.47.0

# NLP / macros
nltk==3.9.1

# OCR
pytesseract==0.3.13
Pillow==10.4.0

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1
